Set up arping in Pulsar Timer.
This was in an attempt to get Tom's Hebe machine to read packets. Added some commented-out lines as well to
imitate tut2 but this turned out to only work with a newer version of casperfpga, and I don't feel like wading
into that particular minefield just yet.

Turns out that this didn't make a difference, the problem was somewhere else. But I'm leaving it in, as it's
good practice.

Changes to be committed:
  modified:   LaunchPulsarTimer.py