Changed logic to use _17 instead of _0
Corresponds to commit 5effcb5 on AVNRoachGateware repo.
I'm hoping that this will help a bit with the linearity and with the dynamic range.
I'm concerned about the accumulators though. I did change them to have fractional
bits and to round instead of truncate, but I'll see what happens.

Changes to be committed:
	modified:   nb_spectrometer.bof
	modified:   nb_spectrometer.fpg
	modified:   wb_spectrometer.bof
	modified:   wb_spectrometer.fpg