Further timing corrections, and launch script updates.
Corresponds to commit 285b10 on the AVNRoachGateware repo.
In the WB spectrometer case, the alignment of the acc_trig pulse needed two further
clock delays in order to line up.

In the NB spectrometer case, the requant after digital gain in Stokes block
has latency of 4, and I forgot to add delays to the DV and the Sync lines. Corrected
now.

I also modified the Launcher script to always clobber the bof file, doesn't even
bother to check whether it's in the /srv/roachfs/fs/boffiles directory first. This
just reducees accidental errors when pulling a new repo version.